code/Backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.1.1
flask_login
flask_wtf
wtforms
wtforms.validators
regex
python-dotenv
bcrypt
flask_marshmallow
marshmallow-sqlalchemy
oauthlib
authlib
pyjwt
requests
google.auth
email_validator
selenium
Werkzeug==2.3.0
gunicorn